% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{The bundle S7 class}
\usage{
bundle(streamlines = list(), bundle_data = list())
}
\arguments{
\item{streamlines}{A list of \link{streamline} objects.}

\item{bundle_data}{A named list of bundle-level metadata.}
}
\description{
A \link{bundle} is an ordered collection of \link{streamline} objects representing a
tractogram or white-matter bundle.  It stores two compartments:
\itemize{
\item \verb{@streamlines} — a list of \link{streamline} objects.
\item \verb{@bundle_data} — a named list of bundle-level metadata (arbitrary R
objects, e.g. the affine transform used during tracking).
}
}
\details{
Use the \code{\link[=new_bundle]{new_bundle()}} constructor to create instances.
}
