% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dinv}
\alias{dinv}
\title{Vector inverse and diagonal function}
\usage{
dinv(x)
}
\arguments{
\item{x}{the vector}
}
\description{
Calculate the inverse of a vector (1 / x) and set
non-finite values back to 1 and diagonalize it
}

