% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dk.2011}
\alias{dk.2011}
\title{Danish national accounts data}
\format{A list with data on sector-level: gross output (xbar,
    vector), intermediate consumption (ubar, vector). Additionally
    a vector "c0" with the absolute values of the negative entries
    of the trade and transport margins (see SUTRAS paper), as well
    as total imports.}
\usage{
dk.2011
}
\description{
Danish national accounts data, aka projection data that is used to
update the SUT.
}
\references{
\url{http://ec.europa.eu/eurostat/data/database}
}
\author{
Oliver Reiter \email{oliver.reiter@snapdragon.cc}
}
\keyword{data}
