% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doGRAS.R
\name{doGRAS}
\alias{doGRAS}
\title{Do the GRAS algorithm for updating matrizes}
\usage{
doGRAS(A, u, v, epsilon = 0.0000000001, verbose = FALSE)
}
\arguments{
\item{A}{the "base" matrix}

\item{u}{vector with the row totals}

\item{v}{vector with the column totals}

\item{epsilon}{the error tolerance level, default is 1e-10.}

\item{verbose}{should some information of the iterations be
displayed? Default is FALSE}
}
\value{
the updated matrix X
}
\description{
This function calculates an updated matrix X, based on a matrix A,
which meets the given row and columns totals. Note: A can contain
negative elements.
}
\examples{
## example from the papers
A <- matrix(c(7, 3, 5, -3, 2, 9, 8, 0, -2, 0, 2, 0),
            ncol = 4, nrow = 3, byrow = TRUE)
u <- c(15, 26, -1)
v <- c(9, 16, 17, -2)
doGRAS(A, u, v)
}
\author{
Oliver Reiter
}
\references{
Junius T. and J. Oosterhaven (2003), The solution of
    updating or regionalizing a matrix with both positive and
    negative entries, Economic Systems Research, 15, pp. 87-96.
    Lenzen M., R. Wood and B. Gallego (2007), Some comments on the
    GRAS method, Economic Systems Research, 19, pp. 461-465.
    Temurshoev, U., R.E. Miller and M.C. Bouwmeester (2013), A note
    on the GRAS method, Economic Systems Research, 25, pp. 361-367.
}
\keyword{gras,}
\keyword{matrix}
\keyword{updating}

