% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSUTRAS_DomImp.R
\name{doSUTRAS.DomImp}
\alias{doSUTRAS.DomImp}
\title{do a SUTRAS update}
\usage{
doSUTRAS.DomImp(V, Ud, Um, m0, ubar, xbar, M, c0, epsilon = 0.000001,
  max.iterations = 10000, verbose = FALSE)
}
\arguments{
\item{V}{the supply matrix}

\item{Ud}{the domestic use matrix}

\item{Um}{the import use matrix}

\item{m0}{the import vector (as normally included in the supply
matrix)}

\item{ubar}{intermediate consumption by industry sector}

\item{xbar}{gross output by industry sector}

\item{M}{total imports}

\item{c0}{the absolute (negative) values of the trade margins
vector, see references for details.}

\item{epsilon}{the convergence tolerance, default = 1e-6.}

\item{max.iterations}{the maximal number of iterations that would
be carried out}

\item{verbose}{should diagnostic output of the iterations be
displayed? Default = FALSE.}
}
\description{
do a SUTRAS update when the use table is separated in domestic and
imported products. Row and column totals are met.
}

