% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doSUTRAS}
\alias{doSUTRAS}
\title{Do the SUTRAS algorithm for updating supply and use tables}
\usage{
doSUTRAS(V, Ud, Um, m0, u_bar, x_bar, M, c, epsilon = 0.0000000001,
  maxiter = 1000L, verbose = FALSE)
}
\arguments{
\item{V}{the supply (make) matrix}

\item{Ud}{the domestic use matrix}

\item{Um}{the import use matrix}

\item{m0}{the import vector}

\item{u_bar}{vector with the new intermediate input sums}

\item{x_bar}{vector with the new gross output sums}

\item{M}{scalar, new total imports}

\item{c}{vector with the negative entries of the trade margins, given in absolute values}

\item{epsilon}{the error tolerance level, default is 1e-10.}

\item{maxiter}{maximum number of iterations, default is 10000.}

\item{verbose}{should some information of the iterations be
displayed? Default is FALSE}
}
\value{
the updated SUTs, a list
}
\description{
This function calculates an updated supply and (domestic and import) use
table (SUTs), based on a given SUT and national accounts data.
}
\examples{
data(TestSUTs_AUT2010)
data(TestNA_AUT2011)
c <- rep(0, length(m0))
res <- doSUTRAS(V = V, Ud = Ud, Um = Um, m0 = m0, u_bar = ubar,
                x_bar = xbar, M = M, c = c, epsilon = 1e-10,
                maxiter = 100, verbose = FALSE)
}
\references{
Temurshoev, U., & Timmer, M. P. (2011). Joint estimation of supply and use tables. Papers in Regional Science, 90(4), 863-882.
}
\author{
Oliver Reiter
}
\keyword{SUTRAS,}
\keyword{input-output}
\keyword{updating}
