% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_noti}
\alias{index_noti}
\title{Return a vector of indices except index i}
\usage{
index_noti(i, nm1)
}
\arguments{
\item{i}{index not to be included}

\item{nm1}{length of the vector}
}
\description{
Return a vector of indices except index i
}
