% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{is_bundle}
\alias{is_bundle}
\title{Check whether an object is a bundle}
\usage{
is_bundle(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is of class \link{bundle}, otherwise \code{FALSE}.
}
\description{
Check whether an object is a bundle
}
