% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{is_streamline}
\alias{is_streamline}
\title{Check whether an object is a streamline}
\usage{
is_streamline(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is of class \verb{<riot::streamline>}, otherwise \code{FALSE}.
}
\description{
Check whether an object is a streamline
}
