% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConsistent.R
\name{isConsistent}
\alias{isConsistent}
\title{Function to check the consistency of a SUT system}
\usage{
isConsistent(SUT, tol = 0.0001, VA.included = FALSE)
}
\arguments{
\item{SUT}{the SUT system to be tested}

\item{tol}{the tolerance (epsilon), default = 1e-4}

\item{VA.included}{is a value-added included in the SUT system? If
yes, then the function also checks for consistent
columns. Default = FALSE.}
}
\value{
If the SUT is consistent, returns only TRUE. If not,
    returns FALSE and a list of inconsistent rows and columns
}
\description{
Given a SUT system containing a supply table and a domestic and a
import use table, this function checks if the rows and columns are
consistent.
}
