% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{n_points}
\alias{n_points}
\title{Number of points in a streamline}
\usage{
n_points(x)
}
\arguments{
\item{x}{A \link{streamline} object.}
}
\value{
Integer scalar giving the number of points.
}
\description{
Number of points in a streamline
}
\keyword{internal}
