% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{new_bundle}
\alias{new_bundle}
\title{Create a bundle object}
\usage{
new_bundle(streamlines)
}
\arguments{
\item{streamlines}{A list of objects of class \code{streamline}.}
}
\value{
An object of class \code{bundle}.
}
\description{
A \code{bundle} is an ordered list of \link[=new_streamline]{streamline} objects
representing a collection of fibre tracts (a tractogram or white-matter
bundle).
}
