% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fibers.R
\name{read_fibers}
\alias{read_fibers}
\title{Import fibers into R}
\usage{
read_fibers(file)
}
\arguments{
\item{file}{A path string to the file from tractography algorithm. Currently
supported files are \code{.vtk} and \code{.vtp} files.}
}
\value{
A \code{\link[tibble]{tibble}} storing the set of tracts.
}
\description{
Import fibers into R
}
\examples{
uf_left  <- read_fibers(system.file("extdata", "UF_left.vtp", package = "riot"))
uf_right <- read_fibers(system.file("extdata", "UF_right.vtp", package = "riot"))
}
