% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tractogram.R
\name{read_tractogram}
\alias{read_tractogram}
\title{Import tractograms into R}
\usage{
read_tractogram(file, reference_file = NULL)
}
\arguments{
\item{file}{A string specifying the path to the file containing the tractography data. Currently
supported files are:
\itemize{
\item standard \href{https://vtk.org}{VTK} formats \code{.vtk} and \code{.vtp},
\item \href{https://med.inria.fr}{medInria} \code{.fds} format,
\item \href{https://mrtrix.readthedocs.io/en/latest/getting_started/image_data.html}{MRtrix} \code{.tck/.tsf} format,
\item \href{http://trackvis.org/docs/?subsect=fileformat}{TrackVis} \code{.trk} and \code{.trx} formats,
\item \href{https://docs.dipy.org/1.11.0/}{DIPY} \code{.dpy} format,
\item \code{.fib} format.
}}

\item{reference_file}{A string specifying the path to a reference image file. This is only required
when importing \code{.trx}, \code{.fib}, or \code{.dpy} files, as these formats do not contain spatial
information about the image space. The reference image is used to correctly position the
tractogram in the appropriate space. Default is \code{NULL}.}
}
\value{
A special \link[tibble:tibble]{tibble} (of class \code{maf_df}) containing the
tractogram data with the following columns:
\itemize{
\item \code{X}: X coordinate of the point.
\item \code{Y}: Y coordinate of the point.
\item \code{Z}: Z coordinate of the point.
\item \code{PointId}: Identifier of the point within its streamline.
\item \code{StreamlineId}: Identifier of the streamline.
}

The tractogram may contain additional columns that are attributes to either points or streamlines.
}
\description{
This is the go-to function to import tractograms into R. Based on both VTK and DIPY, we currently
support eight different formats detailed in the documentation of input argument \code{file}.
}
\examples{
uf_left_vtk <- read_tractogram(system.file("extdata", "UF_left.vtk",  package = "riot"))
}
\seealso{
\code{\link[=write_tractogram]{write_tractogram()}} to export tractograms from R.
}
