% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackReaders.R
\name{read_tracts}
\alias{read_tracts}
\title{Read tracts in VTP format}
\usage{
read_tracts(file)
}
\arguments{
\item{file}{A path string to the file from tractography algorithm. Currently
supported files are \code{.vtk} and \code{.vtp} files.}
}
\value{
A \code{\link[tibble]{tibble}} storing the set of tracts.
}
\description{
Read tracts in VTP format
}
\examples{
uf_left <- trio::read_tracts(system.file("extdata", "UF_left.vtp", package = "trio"))
uf_right <- trio::read_tracts(system.file("extdata", "UF_right.vtp", package = "trio"))
}
