% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackReaders.R
\name{read_vtk}
\alias{read_vtk}
\title{Read tracts in VTK format}
\usage{
read_vtk(filename)
}
\arguments{
\item{filename}{A path string to the \code{.vtk} file.}
}
\value{
A \code{\link[tibble]{tibble}} storing the set of tracts.
}
\description{
Read tracts in VTK format
}
