% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackReaders.R
\name{read_vtp}
\alias{read_vtp}
\title{Read tracts in .VTP format}
\usage{
read_vtp(file)
}
\arguments{
\item{file}{A path string to the `.vtp` file.}
}
\value{
A \code{\link[tibble]{tibble}} storing the set of tracts.
}
\description{
Read tracts in .VTP format
}
