% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riot.R
\docType{package}
\name{riot}
\alias{riot}
\alias{riot-package}
\title{riot}
\description{
R Input-Output Tools contains utility functions for input-output database related stuff. Input-Output as the matrizes Leontief invented.
}
\author{
Oliver Reiter
}
