% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwCentrality}
\alias{rwCentrality}
\title{Calculate random walk centrality}
\usage{
rwCentrality(A, verbose = FALSE)
}
\arguments{
\item{A}{the input-output matrix}

\item{verbose}{should some information of the iterations be
displayed? Default is FALSE}
}
\value{
a vector of centralities
}
\description{
Calculate random walk centrality of an input-output matrix
}
\references{
Blöchl F, Theis FJ, Vega-Redondo F, and Fisher E: Vertex Centralities in Input-Output Networks Reveal the Structure of Modern Economies, Physical Review E, 83(4):046127, 2011
}
\author{
Oliver Reiter
}
\keyword{analysis}
\keyword{input-output}
