% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ShermanUpdateCol}
\alias{ShermanUpdateCol}
\title{Update the inverse by column}
\usage{
ShermanUpdateCol(A, u, i)
}
\arguments{
\item{A}{an inverse of a matrix to be updated}

\item{u}{column vector}

\item{i}{row index}
}
\description{
Update the inverse by column
}
