% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ShermanUpdateRow}
\alias{ShermanUpdateRow}
\title{Update the inverse by row}
\usage{
ShermanUpdateRow(A, v, i)
}
\arguments{
\item{A}{an inverse of a matrix to be updated}

\item{v}{row vector}

\item{i}{column index}
}
\description{
Update the inverse by row
}
