% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{[.bundle}
\alias{[.bundle}
\title{Subset a bundle}
\usage{
\method{[}{bundle}(x, i)
}
\arguments{
\item{x}{A \link{bundle} object.}

\item{i}{Integer, logical, or character index vector.}
}
\value{
A \link{bundle} containing the selected streamlines (preserving
\code{bundle_data}).
}
\description{
Subset a bundle
}
