% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{[[.riot::bundle}
\alias{[[.riot::bundle}
\title{Extract a streamline from a bundle by index}
\usage{
\method{[[}{`riot::bundle`}(x, i)
}
\arguments{
\item{x}{A \link{bundle} object.}

\item{i}{Integer or character index.}
}
\value{
The corresponding \link{streamline} object.
}
\description{
Extract a streamline from a bundle by index
}
