% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trio-package.R
\docType{package}
\name{trio-package}
\alias{trio}
\alias{trio-package}
\title{trio: trio: TRactography Input/Output}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/astamm/trio}
  \item Report bugs at \url{https://github.com/astamm/trio/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Benoit Combès
  \item Pascale Fournier
}

}
