% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{V}
\alias{V}
\title{Austria SUTs}
\format{A supply table (V, a matrix) and a separate
    import vector (m0), domestic use table (Ud, matrix) and import
    use table (Um, matrix). Furthermore, the country and time of
    the SUT system.}
\description{
Austrian supply and use tables. The use table is split into domestic
and import use.
}
\references{
\url{http://ec.europa.eu/eurostat/data/database}
}
\author{
Oliver Reiter \email{oliver.reiter@snapdragon.cc}
}
\keyword{data}
