% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-tractogram.R
\name{write_tractogram}
\alias{write_tractogram}
\title{Export tractograms from R}
\usage{
write_tractogram(x, file, reference_file = NULL)
}
\arguments{
\item{x}{An object of class \code{maf_df} storing a tractogram.}

\item{file}{A path to a file into which the tractogram should be saved.
Currently supported files are:
\itemize{
\item standard \href{https://vtk.org}{VTK} formats \code{.vtk} and \code{.vtp},
\item \href{https://med.inria.fr}{medInria} \code{.fds} format,
\item \href{https://mrtrix.readthedocs.io/en/latest/getting_started/image_data.html}{MRtrix} \code{.tck/.tsf} format,
\item \href{http://trackvis.org/docs/?subsect=fileformat}{TrackVis} \code{.trk} and \code{.trx} formats,
\item \href{https://docs.dipy.org/1.11.0/}{DIPY} \code{.dpy} format,
\item \code{.fib} format.
}

Warning: we rely on DIPY to provide support to save tractograms in \code{.trk}, \code{.trx}, \code{.tck}, \code{.dpy} and
\code{.fib} formats. Among these formats, only \code{.trk} and \code{.trx} formats are able to keep track of additional
attributes assigned to either streamlines or points.}
}
\value{
The input tractogram (invisibly) so that the function can be
used in pipes.
}
\description{
Export tractograms from R
}
\examples{
uf_left  <- read_tractogram(system.file("extdata", "UF_left.vtp",  package = "riot"))
\dontrun{
out <- fs::file_temp(ext = ".vtp")
write_tractogram(uf_left, file = out)
}
}
\seealso{
\code{\link[=read_tractogram]{read_tractogram()}} to import tractograms into R.
}
