% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{xbar}
\alias{xbar}
\title{Austrian national accounts data}
\format{Data on sector-level: gross output (xbar,
    vector), intermediate consumption (ubar, vector). Additionally
    there is a scalar M holding total imports.}
\description{
Austrian national accounts data, aka projection data that is used to
update the SUT.
}
\references{
\url{http://ec.europa.eu/eurostat/data/database}
}
\author{
Oliver Reiter \email{oliver.reiter@snapdragon.cc}
}
\keyword{data}
