% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSUTRAS_DomImp.R
\name{doSUTRAS.DomImp}
\alias{doSUTRAS.DomImp}
\title{do a SUTRAS update}
\usage{
doSUTRAS.DomImp(SUT, ProjData, epsilon = 0.000001, max.iterations = 10000,
  verbose = FALSE)
}
\arguments{
\item{SUT}{a SUT object containing supply and use matrizes}

\item{ProjData}{Projection (forecast) data containing intermediate
consumption and gross output}

\item{epsilon}{the convergence tolerance, default = 1e-6.}

\item{max.iterations}{the maximal number of iterations that would
be carried out}

\item{verbose}{should diagnostic output of the iterations be
displayed? Default = FALSE.}
}
\description{
do a SUTRAS update when the use table is separated in domestic and
imported products. Row and column totals are met.
}

