% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{new_bundle}
\alias{new_bundle}
\title{Create a bundle object}
\usage{
new_bundle(streamlines, bundle_data = list())
}
\arguments{
\item{streamlines}{A list of \link{streamline} objects.}

\item{bundle_data}{A named list of bundle-level metadata.  Defaults to an
empty list.}
}
\value{
An object of class \link{bundle}.
}
\description{
A convenience constructor that wraps the \link{bundle} S7 class.
}
\seealso{
\code{\link[=new_streamline]{new_streamline()}}
}
