% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-fascicles.R
\name{read_fascicles}
\alias{read_fascicles}
\title{Import fascicles into R}
\usage{
read_fascicles(file)
}
\arguments{
\item{file}{Path to the file containing the tractography data. Currently
supported files are \code{.vtk} and \code{.vtp} files.}
}
\value{
A \code{\link[tibble]{tibble}} storing the set of fascicles.
}
\description{
Import fascicles into R
}
\examples{
uf_left  <- read_fascicles(system.file("extdata", "UF_left.vtp",  package = "riot"))
uf_right <- read_fascicles(system.file("extdata", "UF_right.vtp", package = "riot"))
}
